# firmware/docs/openapi.yml
openapi: 3.1.0
info:
  title: Project Verdify MVP API
  version: "2.0"
  description: >
    REST API for AI-powered greenhouse MVP. HTTPS required for all endpoints. Metric units on wire.
    ETag used for config/plan. Device identity via device_name (verdify-aabbcc).
    Rate limiting applies to telemetry endpoints to prevent DoS attacks.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.verdify.ai/api/v1
...
security:
  - UserJWT: []
paths:
...
  /crops:
    get:
      operationId: listCrops
      tags: [CRUD]
      summary: List crops
      security:
        - UserJWT: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: List of crops with pagination
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CropList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createCrop
      tags: [CRUD]
      summary: Create crop
      security:
        - UserJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CropCreate' }
      responses:
        '201':
          description: Crop created successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Crop' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /crops/{id}:
    get:
      operationId: getCrop
      tags: [CRUD]
      summary: Get crop
      security:
        - UserJWT: []
      parameters:
        - $ref: '#/components/parameters/CropIdParam'
      responses:
        '200':
          description: Crop details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Crop' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateCrop
      tags: [CRUD]
      summary: Update crop
      security:
        - UserJWT: []
      parameters:
        - $ref: '#/components/parameters/CropIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CropUpdate' }
      responses:
        '200':
          description: Crop updated successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Crop' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteCrop
      tags: [CRUD]
      summary: Delete crop
      security:
        - UserJWT: []
      parameters:
        - $ref: '#/components/parameters/CropIdParam'
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    UserJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User JWT for app/admin calls
    DeviceToken:
      type: apiKey
      in: header
      name: X-Device-Token
      description: Long-lived device token for controllers
  parameters:
...
    CropIdParam:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
    PageParam:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      in: query
      name: page_size
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    BadRequest:
      description: Bad request
      headers:
        X-Request-Id:
          description: Request identifier for tracing
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            ex:
              value:
                error_code: E400_BAD_REQUEST
                message: "Invalid input"
                timestamp: "2025-08-13T18:05:00Z"
                details:
                  field: "email"
                  reason: "Invalid email format"
                request_id: "req_1234567890abcdef"
    Unauthorized:
      description: Unauthorized
      headers:
        X-Request-Id:
          description: Request identifier for tracing
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            ex:
              value:
                error_code: E401_UNAUTHORIZED
                message: "Missing/invalid token"
                timestamp: "2025-08-13T18:05:00Z"
                request_id: "req_1234567890abcdef"
    NotFound:
      description: Not found
      headers:
        X-Request-Id:
          description: Request identifier for tracing
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            ex:
              value:
                error_code: E404_NOT_FOUND
                message: "Resource not found"
                timestamp: "2025-08-13T18:05:00Z"
                request_id: "req_1234567890"
...
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - E400_BAD_REQUEST
        - E401_UNAUTHORIZED
        - E403_FORBIDDEN
        - E404_NOT_FOUND
        - E409_CONFLICT
        - E412_PRECONDITION_FAILED
        - E422_UNPROCESSABLE
        - E429_TOO_MANY_REQUESTS
        - E500_INTERNAL
        - E503_SERVICE_UNAVAILABLE
      description: Standardized error codes used across the API
...
    ErrorResponse:
      type: object
      required: [error_code, message, timestamp, request_id]
      properties:
        error_code:
          $ref: '#/components/schemas/ErrorCode'
          description: Standardized error code
          example: "E409_CONFLICT"
        message:
          type: string
          description: Human-readable error message
          example: "Email already exists"
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the error occurred
        request_id:
          type: string
          description: Unique request identifier for tracing
          example: "req_1234567890"
        details:
          oneOf:
            - type: array
              items: { $ref: '#/components/schemas/ErrorDetail' }
            - type: object
              additionalProperties: true
            - type: string
          description: Structured error details for field-specific issues
          example:
            field: "email"
            value: "user@example.com"
    ErrorDetail:
      type: object
      required: [field, value, expected]
      properties:
        field:
          type: string
          description: Name of the field that caused the error
          example: "email"
        value:
          description: The actual value that was provided
          example: "invalid-email"
        expected:
          type: string
          description: What was expected for this field
          example: "Valid email address"

    # --- Authentication ---
...
    Crop:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 255 }
        description: { type: [string, "null"], maxLength: 500 }
        expected_yield_per_sqm: { type: [number, "null"] }
        growing_days: { type: [integer, "null"] }
        recipe:
          anyOf:
            - type: object
              additionalProperties: true
            - type: "null"
    CropCreate:
      required: [name]
      type: object
      properties:
        name: { type: string }
        description: { type: [string, "null"] }
        expected_yield_per_sqm: { type: [number, "null"] }
        growing_days: { type: [integer, "null"] }
        recipe:
          anyOf:
            - type: object
              additionalProperties: true
            - type: "null"
    CropUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: [string, "null"] }
        expected_yield_per_sqm: { type: [number, "null"] }
        growing_days: { type: [integer, "null"] }
        recipe:
          anyOf:
            - type: object
              additionalProperties: true
            - type: "null"
    CropList:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Crop' }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
...